requirements.txt
pandas>=2.2.0
numpy>=1.24.0
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyYAML>=6.0
loguru>=0.7.2
streamlit>=1.28.0
PyPDF2>=3.0.1
